Replace ServiceKeySpec with ServiceKeyTests.
